requirements.txt
flask==2.3.3
google-cloud-storage==2.13.0
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
google-auth==2.23.3